requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
python-dotenv==1.0.0

# QR Code & PDF Generation
qrcode[pil]==7.4.2
Pillow==10.2.0
reportlab==4.0.9

# Email & SMS (Optional)
python-decouple==3.8

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
